Add a Github workflow to build and deploy the nar
The workflow is triggered when a release tag is pushed. We use
different jobs to build for different platforms (macOS, Linux,
Windows), and a fourth job consolidates everything, builds the nar and
deploys to Sonatype. A macOS runner is used to build for macOS, Linux
runners are used to build for Linux and Windows. We still use
cross-compiling for the Windows build. We don't need Docker. An option
is provided to trigger the workflow manually.

See PR https://github.com/liblouis/liblouis-nar/pull/6.